use-cases/02-rag-enterprise/requirements.txt
anthropic>=0.84.0
fastapi>=0.135.0
uvicorn>=0.41.0
pydantic>=2.12.0
python-dotenv>=1.2.0
chromadb>=0.6.0
sentence-transformers>=3.0.0
python-multipart>=0.0.20
pypdf>=5.0.0
httpx>=0.28.0
